Redesigned the S3 Cache test to utilize a public external resource and use public cache apis
